Vectorize smartbar buttons popup selected item
PS2 - Vector all things! No need to keep using a .9.png image that
will stay the same color until a themer changes when you can use
a vector and add the accent color to it. The advantage here if is
not obvious is that this will change with themes automatically if
the themer for one reason or another doesn't change it.

PS3 - Change accent to use the darker variant so you can still see the menu
icon.

Change-Id: I80683e4055aac321775f7d1138783fc54948f8d6
Signed-off-by: spezi77 <spezi77@gmx.com>