Fix smartbar editor edit popups not showing sometimes
e.g. when system animations are set to 0.5x in dev settings,
or when triggering the popup a few times, and other potential
cases.

PopupWindow let us set a new popup content without dismissing the
existing popup, so we can get rid of the -sometimes broken- dismiss
listener to show the new content for the 2nd popup

Signed-off-by: spezi77 <spezi77@gmx.com>